% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearBaseline.R
\name{LinearBaseline}
\alias{LinearBaseline}
\title{LinearBaseline}
\usage{
LinearBaseline(x)
}
\arguments{
\item{x}{data matrix with rows as samples and columns as variables}

\item{prd.td}{cutoff parameter}
}
\value{
The function returns object of type "matrix".
\item{x}{  matrix after normalization}
}
\description{
Linear Baseline normalization.
}
\details{
Linear Baseline normalization as described in [1].
}
\examples{
data(mcad)
x <- mcad[,-1]
LinearBaseline(x)
}
\references{
[1] ADD NAME OF BOOK AND CHAPTER
}
\author{
Jan Walach

Jan Walach
}
