% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonLinearBaseline.R
\name{NonLinearBaseline}
\alias{NonLinearBaseline}
\title{NonLinearBaseline}
\usage{
NonLinearBaseline(x, prd.td = c(0.003, 0.007))
}
\arguments{
\item{x}{data matrix with rows as samples and columns as variables}

\item{prd.td}{cutoff parameter}
}
\value{
The function returns object of type "matrix".
\item{liwong.x}{  matrix after normalization}
}
\description{
Non-Linear Baseline normalization.
}
\details{
Non-Linear Baseline normalization as described in [1]. The function is based on function from Affy package [2]
}
\examples{
data(mcad)
x <- mcad[,-1]
NonLinearBaseline(x,prd.td=c(0.003,0.007))
}
\references{
[1] CAC Vol 82: Data Analysis for Omic Sciences: Methods and Applications chapters are due, Chapter Data normalization and scaling:  consequences forthe analysis in omics sciences

[2] L.Gautier and L. Cope and B.M. Bolstad and R. A. Irizarry (2004) affy---analysis of Affymetrix GeneChip data at the probe level, Bioinformatics, 20(3), 307--315, 2004
}
\author{
Jan Walach

Jan Walach
}
